separating out command tests
